+++ engines/dfc_provider/app/controllers/dfc_provider/supplied_products_controller.rb
@@ -1,0 +1,20 @@
+# frozen_string_literal: true
+
+# Controller used to provide the SuppliedProducts API for the DFC application
+# SuppliedProducts are products that are managed by an enterprise.
+module DfcProvider
+  class SuppliedProductsController < DfcProvider::BaseController
+    before_action :check_enterprise
+
+    def show
+      render json: variant, serializer: DfcProvider::SuppliedProductSerializer
+    end
+
+    private
+
+    def variant
+      @variant ||=
+        DfcProvider::VariantFetcher.new(current_enterprise).scope.find(params[:id])
+    end
+  end
+end
