+++ engines/dfc_provider/app/serializers/dfc_provider/person_serializer.rb
@@ -26,7 +26,7 @@ module DfcProvider
     end
 
     def id
-      dfc_provider_routes.api_dfc_provider_person_url(
+      dfc_provider_routes.dfc_provider_person_url(
         id: object.id,
         host: host
       )
