+++ engines/dfc_provider/spec/controllers/dfc_provider/catalog_items_controller_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe DfcProvider::Api::CatalogItemsController, type: :controller do
+describe DfcProvider::CatalogItemsController, type: :controller do
   render_views
 
   let!(:user) { create(:user) }
