+++ engines/dfc_provider/spec/controllers/dfc_provider/enterprises_spec.rb
@@ -2,13 +2,12 @@
 
 require 'spec_helper'
 
-describe DfcProvider::Api::SuppliedProductsController, type: :controller do
+describe DfcProvider::EnterprisesController, type: :controller do
   render_views
 
   let!(:user) { create(:user) }
   let!(:enterprise) { create(:distributor_enterprise, owner: user) }
   let!(:product) { create(:simple_product, supplier: enterprise ) }
-  let!(:variant) { product.variants.first }
 
   describe '.show' do
     context 'with authorization token' do
@@ -25,21 +24,24 @@ describe DfcProvider::Api::SuppliedProductsController, type: :controller do
 
         context 'with an enterprise' do
           context 'given with an id' do
-            before do
-              api_get :show, enterprise_id: 'default', id: variant.id
-            end
+            before { api_get :show, id: 'default' }
 
             it 'is successful' do
               expect(response).to be_successful
             end
 
             it 'renders the required content' do
-              expect(response.body).to include(variant.name)
+              expect(response.body)
+                .to include(product.name)
+              expect(response.body)
+                .to include(product.sku)
+              expect(response.body)
+                .to include("offers/#{product.variants.first.id}")
             end
           end
 
           context 'given with a wrong id' do
-            before { api_get :show, enterprise_id: 'default', id: 999 }
+            before { api_get :show, id: 999 }
 
             it 'is not found' do
               expect(response).to be_not_found
