+++ engines/dfc_provider/spec/controllers/dfc_provider/persons_controller_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe DfcProvider::Api::PersonsController, type: :controller do
+describe DfcProvider::PersonsController, type: :controller do
   render_views
 
   let!(:user) { create(:user) }
